pyproject.toml
[project]
name = "pitchpro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.121.0",
    "google-generativeai>=0.8.5",
    "matplotlib>=3.10.7",
    "openai>=2.7.1",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "python-pptx>=1.0.2",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "rq>=2.6.0",
    "streamlit>=1.51.0",
    "uvicorn[standard]>=0.38.0",
]
